changed the backface culling threshold to 0.1 as this is giving better results.
This is probably due to an error in my code, or perhaps a rounding error, as the threshold for visible faces should be 0. However, if I use 0 then there is an odd 'blinking' effect where some sections fo the mesh don't disappear in time and instead linger for a fraction of a second.